Moved default analyzer generators to new file.
